Week 0 Problem Set: Dress Me Up
This is a 'Dress Me Up' styling game. When I was younger, all I played were dressed up games and even wanted to be a fashion designer. The beauty of coding is that I can program any occupation as a game! Hence, I decided to pursue computer science for university. 

How to play?
The instructions comes with the game already.